Add get collection test
